+++ tests/auth-anonymous.test.js
@@ -59,62 +59,6 @@ describe('Auth for anonymous users', () => {
     });
   });
 
-  describe('Mutation.logout', () => {
-    it('log out userthatlogsout', async () => {
-      const Users = db.collection('users');
-      await Users.findOrInsertOne({
-        ...User,
-        _id: 'userthatlogsout',
-        username: 'userthatlogsout',
-        emails: [
-          {
-            address: 'userthatlogsout@localhost',
-            verified: true
-          }
-        ],
-        services: {
-          ...User.services,
-          resume: {
-            loginTokens: [
-              {
-                when: new Date(new Date().getTime() + 1000000),
-                hashedToken: 'dF4ilYpWpsSvkb7hdZKqsiYa207t2HI+C+HJcowykZk='
-              }
-            ]
-          }
-        }
-      });
-      const { data: { logout } = {} } = await graphqlFetch({
-        query: /* GraphQL */ `
-          mutation logout($token: String!) {
-            logout(token: $token) {
-              success
-            }
-          }
-        `,
-        variables: {
-          token: 'FWhglvqdNkNX80jZMJ61FvDUkKzCsESVfbui9H8Fg27'
-        }
-      });
-      expect(logout).toMatchObject({
-        success: true
-      });
-    });
-    it('token is gone', async () => {
-      // Reset the password with that token
-      const Users = db.collection('users');
-      const user = await Users.findOne({
-        _id: 'userthatlogsout'
-      });
-      const {
-        services: {
-          resume: { loginTokens }
-        }
-      } = user;
-      expect(loginTokens.length).toEqual(1);
-    });
-  });
-
   describe('Mutation.loginAsGuest', () => {
     it('login as guest', async () => {
       const { data: { loginAsGuest } = {} } = await graphqlFetch({
@@ -227,6 +171,48 @@ describe('Auth for anonymous users', () => {
         }
       });
       expect(resetPassword).toMatchObject({
+        id: 'userthatforgetspasswords',
+        user: {
+          _id: 'userthatforgetspasswords'
+        }
+      });
+    });
+  });
+
+  describe('Mutation.verifyEmail', () => {
+    it('verifies the e-mail of the guest', async () => {
+      // Reset the password with that token
+      const Users = db.collection('users');
+      const guest = await Users.findOne({
+        guest: true
+      });
+      console.log(guest);
+      const {
+        services: {
+          password: {
+            reset: { token }
+          }
+        }
+      } = user;
+
+      const { data: { verifyEmail } = {} } = await graphqlFetch({
+        query: /* GraphQL */ `
+          mutation verifyEmail($newPlainPassword: String, $token: String!) {
+            verifyEmail(newPlainPassword: $newPlainPassword, token: $token) {
+              id
+              token
+              user {
+                _id
+              }
+            }
+          }
+        `,
+        variables: {
+          newPlainPassword: 'password',
+          token
+        }
+      });
+      expect(verifyEmail).toMatchObject({
         id: 'userthatforgetspasswords',
         user: {
           _id: 'userthatforgetspasswords'
