+++ tests/auth-user.test.js
@@ -1,5 +1,5 @@
 import { setupDatabase, createLoggedInGraphqlFetch } from './helpers';
-import { USER_TOKEN } from './seeds/users';
+import { User, USER_TOKEN } from './seeds/users';
 
 let connection;
 let db;
@@ -49,6 +49,62 @@ describe('Auth for logged in users', () => {
       expect(resendVerificationEmail).toMatchObject({
         success: true
       });
+    });
+  });
+
+  describe('Mutation.logout', () => {
+    it('log out userthatlogsout', async () => {
+      const Users = db.collection('users');
+      await Users.findOrInsertOne({
+        ...User,
+        _id: 'userthatlogsout',
+        username: 'userthatlogsout',
+        emails: [
+          {
+            address: 'userthatlogsout@localhost',
+            verified: true
+          }
+        ],
+        services: {
+          ...User.services,
+          resume: {
+            loginTokens: [
+              {
+                when: new Date(new Date().getTime() + 1000000),
+                hashedToken: 'dF4ilYpWpsSvkb7hdZKqsiYa207t2HI+C+HJcowykZk='
+              }
+            ]
+          }
+        }
+      });
+      const { data: { logout } = {} } = await graphqlFetch({
+        query: /* GraphQL */ `
+          mutation logout($token: String!) {
+            logout(token: $token) {
+              success
+            }
+          }
+        `,
+        variables: {
+          token: 'FWhglvqdNkNX80jZMJ61FvDUkKzCsESVfbui9H8Fg27'
+        }
+      });
+      expect(logout).toMatchObject({
+        success: true
+      });
+    });
+    it('token is gone', async () => {
+      // Reset the password with that token
+      const Users = db.collection('users');
+      const user = await Users.findOne({
+        _id: 'userthatlogsout'
+      });
+      const {
+        services: {
+          resume: { loginTokens }
+        }
+      } = user;
+      expect(loginTokens.length).toEqual(1);
     });
   });
 });
